command-center/requirements.txt
Flask==3.1.2
prompt_toolkit>=3.0.0
python-dotenv